+++ megatron/training/training.py
@@ -59,7 +59,6 @@ from megatron.training.initialize import set_jit_fusion_options
 from megatron.training.utils import (
     get_batch_on_this_cp_rank,
     get_batch_on_this_tp_rank,
-    set_manual_pipeline_split_patch,
 )
 from megatron.legacy.data.data_samplers import build_pretraining_data_loader
 from megatron.core.optimizer_param_scheduler import OptimizerParamScheduler
@@ -683,6 +682,7 @@ def pretrain(
 
     # Enable manually split layers in (interleaved) 1f1b pipeline parallelism by monkey patching
     if args.decoder_pipeline_manual_split_list is not None:
+        from megatron.training.patch import set_manual_pipeline_split_patch
         print_rank_0('monkey patch to enable manual pipeline split...')
         set_manual_pipeline_split_patch(args)
